Moved travis webhook into its own file travis.py
