+++ zerver/views/webhooks/__init__.py
@@ -105,55 +105,3 @@ def send_formated_pagerduty(user_profile, stream, message_type, format_dict, top
 
     check_send_message(user_profile, get_client('ZulipPagerDutyWebhook'), 'stream',
                        [stream], subject, body)
-
-
-@api_key_only_webhook_view
-@has_request_variables
-def api_pagerduty_webhook(request, user_profile, stream=REQ(default='pagerduty'), topic=REQ(default=None)):
-    payload = ujson.loads(request.body)
-
-    for message in payload['messages']:
-        message_type = message['type']
-
-        if message_type not in PAGER_DUTY_EVENT_NAMES:
-            send_raw_pagerduty_json(user_profile, stream, message, topic)
-
-        try:
-            format_dict = build_pagerduty_formatdict(message)
-        except:
-            send_raw_pagerduty_json(user_profile, stream, message, topic)
-        else:
-            send_formated_pagerduty(user_profile, stream, message_type, format_dict, topic)
-
-    return json_success()
-
-@api_key_only_webhook_view
-@has_request_variables
-def api_travis_webhook(request, user_profile, stream=REQ(default='travis'), topic=REQ(default=None)):
-    message = ujson.loads(request.POST['payload'])
-
-    author = message['author_name']
-    message_type = message['status_message']
-    changes = message['compare_url']
-
-    good_status = ['Passed', 'Fixed']
-    bad_status  = ['Failed', 'Broken', 'Still Failing']
-    emoji = ''
-    if message_type in good_status:
-        emoji = ':thumbsup:'
-    elif message_type in bad_status:
-        emoji = ':thumbsdown:'
-    else:
-        emoji = "(No emoji specified for status '%s'.)" % (message_type,)
-
-    build_url = message['build_url']
-
-    template = (
-        u'Author: %s\n'
-        u'Build status: %s %s\n'
-        u'Details: [changes](%s), [build log](%s)')
-
-    body = template % (author, message_type, emoji, changes, build_url)
-
-    check_send_message(user_profile, get_client('ZulipTravisWebhook'), 'stream', [stream], topic, body)
-    return json_success()
