+++ zproject/urls.py
@@ -165,7 +165,7 @@ urlpatterns += patterns('zerver.views',
     url(r'^api/v1/external/freshdesk$',     'webhooks.freshdesk.api_freshdesk_webhook'),
     url(r'^api/v1/external/zendesk$',       'webhooks.zendesk.api_zendesk_webhook'),
     url(r'^api/v1/external/pagerduty$',     'webhooks.api_pagerduty_webhook'),
-    url(r'^api/v1/external/travis$',        'webhooks.api_travis_webhook'),
+    url(r'^api/v1/external/travis$',        'webhooks.travis.api_travis_webhook'),
 
     url(r'^user_uploads/(?P<realm_id>(\d*|unk))/(?P<filename>.*)', 'get_uploaded_file'),
 )
